run `sudo DevToolsSecurity -enable` during provisioning (#86)
Without this Xcode.app might block UI Tests,
waiting for manual input (password). Since Xcode 9.3 it seems this is
the only way to accept the permissions required for iOS UI Tests,
if you only accept it via the GUI that won't persist after an OS reboot,
but running this command will.

Before Xcode 9.3 it was enough to accept this via the GUI only once,
but this seem to be changed in Xcode 9.3 (final).